+++ sdk/storage/storage-queue/src/utils/utils.common.ts
@@ -424,3 +424,25 @@ export function getStorageClientContext(url: string, pipeline: Pipeline): Storag
   (storageClientContext as any).requestContentType = undefined;
   return storageClientContext;
 }
+
+/**
+ * Append a string to URL query.
+ *
+ * @export
+ * @param {string} url Source URL string.
+ * @param {string} queryParts String to be appended to the URL query.
+ * @returns {string} An updated URL string.
+ */
+export function appendToURLQuery(url: string, queryParts: string): string {
+  const urlParsed = URLBuilder.parse(url);
+
+  let query = urlParsed.getQuery();
+  if (query) {
+    query += "&" + queryParts;
+  } else {
+    query = queryParts;
+  }
+
+  urlParsed.setQuery(query);
+  return urlParsed.toString();
+}