+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Updated Azure Storage Service API version to 2020-04-08.
 - Share now supports for NFS. User can specify the `protocols` when creating a Share via `ShareClient.create()`. Also added an `rootSquash` option for NFS shares.
+- Added `generateSasUrl` to `ShareClient` and `ShareFileClient` to generate a service-level SAS URI for the client.
+- Added `generateAccountSasUrl` to `ShareServiceClient` to generate an account-level SAS URI for the client.
 - Won't remove the first space in the `userAgentOptions.userAgentPrefix` passed to the `newPipeline()` now. Fixed bug [7536](https://github.com/Azure/azure-sdk-for-js/issues/7536).
 
 ## 12.3.0 (2020-11-10)